refactor(architecture): resolve 12/15 clean architecture violations — 97% compliance
Phases 2-5 of the refactoring plan:

**Domain enrichment (Phase 2):**
- Move reorder logic to Competition.reorder_golf_courses_phase1/phase2()
- Move update workflow to GolfCourse.apply_update() (clone vs in-place)
- Extract compute_strokes_received() to PlayingHandicapCalculator
- Add ScheduleFormatService for match format sequencing

**DI injection & simplification (Phase 3):**
- Inject SnakeDraftService, PlayingHandicapCalculator, ScheduleFormatService via DI
- Refactor reassign_match_players: extract _build_handicap_data/_build_match_player

**Infrastructure cleanup (Phase 4):**
- Split competition_routes.py (1627 lines) into 3 focused route files
- Move CompetitionDTOMapper to application/mappers/ (correct layer)
- Fix team_assignment column: String(20) → TeamAssignmentModeDecorator
- Centralize 10 Golf Course DI functions to dependencies.py

**Exception centralization (Phase 5):**
- Add MatchNotFoundError, CompetitionNotDraftError, InsufficientPlayersError
- 7 use cases now import from centralized exceptions.py
- Update DDD violations report: 90% → 97% compliance

All 1,331 unit + 222 integration tests passing. Zero ruff errors.
