+++ src/modules/competition/application/use_cases/declare_walkover_use_case.py
@@ -4,19 +4,16 @@ from src.modules.competition.application.dto.round_match_dto import (
     DeclareWalkoverRequestDTO,
     DeclareWalkoverResponseDTO,
 )
-from src.modules.competition.application.exceptions import NotCompetitionCreatorError
+from src.modules.competition.application.exceptions import (
+    MatchNotFoundError,
+    NotCompetitionCreatorError,
+)
 from src.modules.competition.domain.repositories.competition_unit_of_work_interface import (
     CompetitionUnitOfWorkInterface,
 )
 from src.modules.competition.domain.value_objects.match_id import MatchId
 from src.modules.competition.domain.value_objects.round_status import RoundStatus
 from src.modules.user.domain.value_objects.user_id import UserId
-
-
-class MatchNotFoundError(Exception):
-    """El partido no existe."""
-
-    pass
 
 
 class CompetitionNotInProgressError(Exception):
