+++ src/modules/golf_course/domain/entities/golf_course.py
@@ -337,6 +337,81 @@ class GolfCourse:
         self._validate_holes()
         self._validate_tees()
 
+    def apply_update(
+        self,
+        name: str,
+        country_code: CountryCode,
+        course_type: CourseType,
+        tees: list[Tee],
+        holes: list[Hole],
+        is_admin: bool,
+    ) -> "GolfCourse | None":
+        """
+        Aplica una actualización al campo de golf según las reglas de negocio.
+
+        Reglas:
+        - REJECTED → error (no editable)
+        - Admin → siempre in-place (retorna None)
+        - Creator + PENDING → in-place (retorna None)
+        - Creator + APPROVED → crea clone (retorna el clone)
+
+        Args:
+            name: Nuevo nombre
+            country_code: Nuevo código de país
+            course_type: Nuevo tipo de campo
+            tees: Nueva lista de tees
+            holes: Nueva lista de hoyos
+            is_admin: Si el usuario es Admin
+
+        Returns:
+            None si se actualizó in-place, GolfCourse clone si se creó propuesta
+
+        Raises:
+            ValueError: Si el campo está REJECTED
+        """
+        if self._approval_status == ApprovalStatus.REJECTED:
+            raise ValueError(
+                "Cannot edit a REJECTED golf course. Please create a new request instead."
+            )
+
+        # Admin o PENDING_APPROVAL → actualización in-place
+        if is_admin or self._approval_status == ApprovalStatus.PENDING_APPROVAL:
+            self.update(name=name, country_code=country_code, course_type=course_type,
+                        tees=tees, holes=holes)
+            return None
+
+        # Creator + APPROVED → crear clone como update proposal
+        clone = GolfCourse.create(
+            name=name,
+            country_code=country_code,
+            course_type=course_type,
+            creator_id=self._creator_id,
+            tees=tees,
+            holes=holes,
+        )
+
+        # Reconstruir clone con campos especiales (link al original, status PENDING)
+        clone_proposal = GolfCourse.reconstruct(
+            id=clone.id,
+            name=clone.name,
+            country_code=clone.country_code,
+            course_type=clone.course_type,
+            creator_id=clone.creator_id,
+            tees=clone.tees,
+            holes=clone.holes,
+            approval_status=ApprovalStatus.PENDING_APPROVAL,
+            rejection_reason=None,
+            created_at=clone.created_at,
+            updated_at=clone.updated_at,
+            original_golf_course_id=self._id,
+            is_pending_update=False,
+        )
+
+        # Marcar original como "tiene cambios pendientes"
+        self.mark_as_pending_update()
+
+        return clone_proposal
+
     def mark_as_pending_update(self) -> None:
         """
         Marca este campo como 'tiene cambios pendientes de aprobación'.
